+++ src/ray.rs
@@ -20,7 +20,7 @@ impl Ray {
         &self.direction
     }
 
-    pub fn point_at_parameter(&self, t: f64) -> Vec3 {
+    pub fn point_at_parameter(&self, t: f32) -> Vec3 {
 
         self.origin + self.direction * t
     }
