+++ src/vec3.rs
@@ -2,17 +2,17 @@ use std::ops::{Add, Sub, Mul, Div};
 
 #[derive(Copy, Clone, Debug)]
 pub struct Vec3 {
-    pub e: [f64; 3],
+    pub e: [f32; 3],
 }
 
 impl Vec3 {
-    pub fn x(self) -> f64 {self.e[0]}
-    pub fn y(self) -> f64 {self.e[1]}
-    pub fn z(self) -> f64 {self.e[2]}
+    pub fn x(self) -> f32 {self.e[0]}
+    pub fn y(self) -> f32 {self.e[1]}
+    pub fn z(self) -> f32 {self.e[2]}
 
-    pub fn r(self) -> f64 {self.e[0]}
-    pub fn g(self) -> f64 {self.e[1]}
-    pub fn b(self) -> f64 {self.e[2]}
+    pub fn r(self) -> f32 {self.e[0]}
+    pub fn g(self) -> f32 {self.e[1]}
+    pub fn b(self) -> f32 {self.e[2]}
 
 }
 
@@ -24,11 +24,11 @@ impl Vec3 {
         Vec3{e:[0.0, 0.0, 0.0]}
     }
 
-    pub fn length(&self) -> f64 {
+    pub fn length(&self) -> f32 {
         (self.e[0] * self.e[0] + self.e[1] * self.e[1] + self.e[2] * self.e[2]).sqrt()
     }
 
-    pub fn squared_length(&self) -> f64 {
+    pub fn squared_length(&self) -> f32 {
         self.e[0] * self.e[0] + self.e[1] * self.e[1] + self.e[2] * self.e[2]
     }
 }
@@ -63,10 +63,10 @@ impl Mul<Vec3> for Vec3 {
     }
 }
 
-impl Mul<f64> for Vec3 {
+impl Mul<f32> for Vec3 {
     type Output = Vec3;
 
-    fn mul(self, rhs:f64) -> Vec3 {
+    fn mul(self, rhs:f32) -> Vec3 {
         Vec3{e:[self.e[0] * rhs, self.e[1] * rhs, self.e[2] * rhs]}
     }
 }
@@ -81,10 +81,10 @@ impl Div<Vec3> for Vec3 {
     }
 }
 
-impl Div<f64> for Vec3 {
+impl Div<f32> for Vec3 {
     type Output = Vec3;
 
-    fn div(self, rhs:f64) -> Self {
+    fn div(self, rhs:f32) -> Self {
         Vec3{e:[ self.e[0]/rhs, self.e[1] / rhs, self.e[2] / rhs]}
     }
 }
@@ -94,7 +94,7 @@ pub fn unit_vector(v: Vec3) -> Vec3 {
     v/length
 }
 
-pub fn dot(v1: &Vec3, v2: &Vec3) -> f64 {
+pub fn dot(v1: &Vec3, v2: &Vec3) -> f32 {
     v1.e[0] * v2.e[0] + v1.e[1] * v2.e[1] + v1.e[2] * v2.e[2]
 }
